updated stuff.  redoing the prefabs for now to get them looking more like the planned ui.  also got the timmer running.  need to get the bullet counting in order so that the dead bullets take away from the counter.

+++ JDBaconJeweled/Unity/Assets/Scripts/JDBaconJeweled/JDBaconJeweled/LevelProgression/JDLevel.cs
@@ -13,8 +13,8 @@ public class JDLevel : JDIObject
 {
     public string Name { get; set; }
     public int Id { get; set; }
-    public int ZombieCollectRate { get; set; }
-    public int ZombieLimit { get; set; }
+    public float ZombieCollectRate { get; set; }
+    public float ZombieLimit { get; set; }
 
     public JDIObjectTypes JDType { get { return JDIObjectTypes.OBJECT; } }
